api_yamdb/requirements.txt
asgiref==3.2.10
atomicwrites==1.4.1
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.0.12
colorama==0.4.5
Django==2.2.28
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-isort==5.0.0
flake8-plugin-utils==1.3.2
flake8-return==1.2.0
gunicorn==20.0.4
idna==3.3
iniconfig==1.1.1
isort==5.10.1
mccabe==0.7.0
packaging==21.3
pep8-naming==0.13.2
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.21.0
pytz==2022.1
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.11